ref: rm hardcoded style values; dynamic title; fix filename
